initial commit
added clearance
added facebook login using https://github.com/NextAcademy/how_to_integrate_clearance_wtih_facebook

right now there seems to be a bug in the instructions, because it seems in Rails 5
the belongs_to method will enforce the referential integrity of the foreign key.

since the authentication belongs_to :user whenever an authentication is created it has to
have a user associated with it.